modificacoes no influx backend feitas na reunião 11/04

+++ chronos/backends/influx.py
@@ -1,12 +1,39 @@
 """InfluxDB backend"""
-from datetime import datetime
-
 from influxdb import InfluxDBClient
 
-from utils import Utils
+from utils import validate_timestamp, now
 
-def now():
-    return datetime.utcnow().isoformat()
+## telemetry
+#
+#kytos.telemetry.log.error = []
+#kytos.telemetry.log.critical = []
+#
+#kytos.telemetry.switches.1.interfaces.232.bytes.in = (ts, value)
+#
+#kytos.telemetry.switches.1.interfaces.233.bytes_in = 
+#kytos.telemetry.switches.1.interfaces.233.bytes_out = 
+#kytos.telemetry.switches.1.interfaces.233.packets_in = 
+#kytos.telemetry.switches.1.interfaces.233.packets_out = 
+#
+#backend.save(namespace, value, ts):
+#backend.save(namespace, values, ts):
+#    # Caso 1 :
+#    measurement -> kytos.telemetry.switches.1.interfaces.232.bytes.in
+#    fields -> "value" = value
+#    
+#    # Caso 2 :
+#    measurement -> kytos.telemetry.switches.1.interfaces.232.bytes
+#    fields -> "in" = value
+#    
+#    # Caso 3 :
+#    measurement -> kytos.telemetry
+#    fields -> "switch" = 1, "interface" = 232, "bytes_in" = 2000
+#    
+#    # Caso 4 :
+#    measurement -> bytes
+#    tags -> "switch" = 1, "interface" = 232
+#    fields -> "in" = 2000, "out" = 300
+
 
 class InfluxBackend:
     """This Backend is responsible to the connection with InfluxDB."""
@@ -32,48 +59,58 @@ class InfluxBackend:
         self._write_endpoints(data)
 
     def delete(self, namespace, start_timestamp=None, end_timestamp=None):
-        """Delete the entire database --
-        start_timestamp and end_timestamp most be a timestamp"""
-        self._validate_namespace('.'.join(namespace.split('.')[:-1]))
-        utils_obj = Utils()
-        utils_obj.validate_timestamp(start_timestamp, end_timestamp)
+        """Delete the entire database.
+
+        start_timestamp and end_timestamp most be a timestamp
+        """
+        if not self._namespace_exists('.'.join(namespace.split('.')[:-1])):
+            raise Exception("Namespace {} does not exist".format(namespace))
+
+        validate_timestamp(start_timestamp, end_timestamp)
+
         self._delete_points(namespace, start_timestamp, end_timestamp)
 
     def get(self, namespace, start_timestamp=None, end_timestamp=None):
         """Make a query to retrieve something in the database."""
-        self._validate_namespace('.'.join(namespace.split('.')[:-1]))
-        utils_obj = Utils()
-        utils_obj.validate_timestamp(start_timestamp, end_timestamp)
-        self._get_points(namespace, start_timestamp, end_timestamp)
+        if not self._namespace_exists('.'.join(namespace.split('.')[:-1])):
+            return None
+
+        validate_timestamp(start_timestamp, end_timestamp)
+        
+        return self._get_points(namespace, start_timestamp, end_timestamp)
 
     def _read_config(self, settings):
         
-        self.params = {'HOST':None,'PORT':None,'USER':None,'PASS':None,'DBNAME':None}
-        for i in settings.BACKENDS['INFLUXDB']:
-            self.params[i] = settings.BACKENDS['INFLUXDB'][i]
+        params = {'HOST': 'localhost',
+                  'PORT': '8086',
+                  'DBNAME': None,
+                  'USER': None,
+                  'PASS': None}
 
-        exist = list(filter(lambda x: x == None, self.params.values()))
-        if len(exist): # exists None in the list
-            if self.params['DBNAME'] is None:
-                raise Exception("Error. Most specify database name.")
-            
-            if self.params['USER'] is None and self.params['PASS']:
-                raise Exception("Error. Most specify user's login.")
-            elif self.params['PASS'] and self.params['USER']:
-                raise Exception("Error. Most specify user's password.")
+        config = settings.BACKENDS.get('INFLUXDB')
 
+        for key in params:
+            params.set(key, config.get(key, params['key']))
+
+        if not params['DBNAME']:
+            raise Exception("Error. Must specify database name.")
+
+        self._host = params['HOST']
+        self._port = params['PORT']
+        self._username = params['USER']
+        self._password = params['PASS']
+        self._database = params['DBNAME']
+       
     def _start_client(self):
-        self._client = InfluxDBClient(host=self._host,
-                                      port=self._port,
+        self._client = InfluxDBClient(host = self._host,
+                                      port = self._port,
                                       username = self._username,
                                       password = self._password,
                                       database = self._database)
 
     def _create_database(self):
-        try:
-            self._client.create_database(self._db_name)
-        except Exception:
-            print("Error creating database")
+        self._client.create_database(self._database)
+
     def _write_endpoints(self, data, create_database=True):
         if not self._get_database() and create_database:
             self._create_database()
@@ -86,7 +123,7 @@ class InfluxBackend:
             raise Exception("Error inserting data to InfluxDB.")
 
     def _get_database(self):
-        """Verify if a database exists"""
+        """Verify if a database exists."""
         all_dbs = self._client.get_list_database()
         exist = filter(lambda x: x['name'] == self._db_name, all_dbs)
         
@@ -128,12 +165,12 @@ class InfluxBackend:
 
         return result_query
 
-    def _validate_namespace(namespace):
+    def _namespace_exists(self, namespace):
 
         if namespace is None:
             raise Exception("Invalid namespace.")
         else:
             all_nspace = self._client.get_list_measurements()
-            exist = filter(lambda x: x['name'] == self.namespace, all_nspace)
+            exist = filter(lambda x: x['name'] == namespace, all_nspace)
             if not exist:
                 raise Exception("Required namespace does not exist.")
